+++ HISTORY.rst
@@ -24,6 +24,8 @@ dev
 * włącz język zapytań dla modułu redagowania: autorzy, źródła, jednostki, itp.
   (b/n),
 * możliwość eksportu danych wydawnictw ciągłych i zwartych do formatu XLSX (b/n),
+* możliwość autoryzacji użytkowników za pomocą protokołu LDAP (b/n),
+* wstępna konfiguracja za pomocą django-environ (b/n),
 
 202207.1087
 -----------